Pipeline Proof of Concept
Contains the initial code for the Pipeline. No metrics have been implemented yet. Codes
serves as a demonstration of how the pipeline will work. The pipeline is currently built
using identity nodes that just pass results through the graph until it reaches a node with
no children.

Pipeline is built on the Directed Acyclic Graph (DAG) data structure. More checks are
to be implemented to ensure that it is meets the DAG criteria.