Pipelined Visqol
Built up from the proof of concept. Added all the necessary node and
funcionality required to recreate the visqol port using the modular
pipeline. Included is the graph config required to construct it.

Also implemented functionality to write both the regular call graph and
the expanded version(which includes graph node graphs etc.) to .dot
files which can then be converted to diagrams using GraphViz