Graph Visualization + Logging fix
Added support to convert the call graph into a graphviz dot file, which
can then be converted to an image(png/svg/etc.). Each node can be
configurared to have a different style when drawing the graph. This is
done through the draw_options entry in the graph config.

Also introduced the AQPNode and ViSQOLNode to support draw_options
better. Different nodes have different draw styles.

Fixed an issue with logging where it wouldn't log to console. This was
caused by using invalid loggers( from visqol) and for some reason using
them as class attributes. To fix simply changed the call to getLogger
and then removed the attribute LOGGER and replaced with a module level
logger.