Mono Signal Fix + Simplified Configuration
Added support for mono signals, before, due to it being skipped over in
a previous port of ViSQOL, it was left out and using a mono signal
caused ViSQOL to crash.

Refactored the way in which nodes get deserialized from JSON. No longer
need both class + module, instead a single 'type' field is used. The
pipeline identifies all available nodes on startup and performs the
dynamic import using this information and the 'type' field.

Also includes some small structural refactoring of nodes into the
visqol_nodes directory. In future will do more refactoring to get a good
structure for nodes, as currently still feels like they're all over the
place.

Small functionality added to the LoopNode which allows a flattening of
the resulting dictionary to the dictionary above it. Best example is
flattening of the channel loop results during batch execution of visqol.
Simply removes a layer of nesting which wasn't necessary. E.g. instead
of dataset_results -> file_names -> vnsims -> channels, can now be
dataset_results ->file_names -> channels.

As of this commit the pipeline is capable of batch executing visqol on a
set files specified in a csv file and having the results available for
each file at the end of execution. This will be used in the future
where the results will be used to create output tables/graphs.