Encapsulation Node
Added the encapsulation node which is capable of parsing a graph
definition from a file and then executing it as normal during the
execution of the pipeline. Purpose of this type of node is that it can
be given a full definition of ViSQOL from a file, this makes writing the
main graph definition shorter and more readable, especially if you're
not changing the parameters of visqol. It can also be used to better
present the call graph when running different configurations against
each other.