Refactoring + Documentation + More Graphs
Refactored the creation and execution of the pipeline. Previously, operated by building a networkx di-graph from json and then traversing it in a depth first fashion to determine how to traverse the pipeline during execution.
Now, there's no need for a networkx graph, instead, each node is initialized from json without making the edge connections between nodes. Once each node is created, it is then updated with a reference to each of it's child nodes.
Ultimately, it still traverses in a depth first fashion, except it is now independent of networkx and pre-determining the traversal order.

The changes to this also allowed for a simplication of a lot of the logic/code. A node is executed, if it returns a result, then evaluate all of it's children, if it returns None, continue executing nodes on the same level of the pipeline. This functionality allows the SinkNode to exist. It is given a number of results it expects to see and then will only pass the result given to it to it's children when it has seen the number of expected results. This allows for the merging of results from different branches. Example of this is running two versions of visqol and then having both results available afer the sink node.

Removed the GraphNode as part of this refactor, as it can be entirely represented using an EncapsulationNode.

Added documention + type hints to almost every function and class.
Added some example graphs.