MOS Mapping + Minor Refactoring + Documentation + README
Added in a node to perform the mapping of similarity data to MOS.

Updated/fixed some of the graph configs.

Added a README, still subject to change.

Removed the ability to create an expanded diagram of the graph. There is issues with it(connecting nested leaf nodes properly) that will be fixed in it's own commit in future.