run_exmaple.sh and prepend_files.py changes
Modified the prepend_files.py script to not be a pain to use. Previously if it was run more than once/on a file that dataframe that already had the desired prefix it would still prefix each cell

Example
-------
resources/myfile.mp4 -> resources/resources/myfile.mp4

or the more generic

{args.prepend_with}/myfile.mp4 -> {args.prepend_with}/{args.prepend_with}/myfile.mp4

The script has now been changed so that it only prefixes cells that don't already start with the desired prefix

Another small modification involves being able to pass through the names of the reference and test columns so that it's not hardcoded to "Ref_Wave" and "Test_Wave"
prepend_files.py has also been moved to the scripts directory
These changes should make the scipt more usable outside of the example scripts

run_example.sh has also been updated. It now checks the return code of the prepend_files.py script and will only run the pipeline if the exit code is 0. Should save some precious seconds :]