Full Pipeline Visualization
- Added full pipeline visualization. This involved a complete rework of the prior code for visualization.
There is now no dependency on networkx or pydot for this. The visualization produced contains information relating to the
subgraph/clusters created by using Nested Nodes(Encapsulation/Loop Nodes). A dot file is still produced, this is created
manually, without the need for a 3rd party library. If the DOT engine is installed and available from the command line then
the pipeline will generate the corresponding pipeline visualization in png and svg formats, if not then just the DOT file
is produced.

- All graph visualization functionality has been moved to the graphvis.py file

- Introduced the ability to define common graph attributes such as rankdir etc. through the use of the config/dot_config.json file.
This file is expected to be there and by default is empty. Each entry in the json file gets converted into the necessary <key>=<value>
format in the DOT file.

- Added the colour_demo.json pipeline configuration. This is mostly for demonstration purposes. It contains a node of each
type so all the colours (which by default should be colour blind friendly) are used in the pipeline to create a visualization.
The visualization produced will be added to the README.md at some stage.

- Added the NestedNode super class. Used for identifying and colouring EncapsulationNodes and LoopNodes differently. They are
represeneted by a dark grey (I think) to indicate that they are mostly just placeholders for their execution graph.

- Created the constants file to clear up issues with the logger

- Added the '--validate' option to the program arguments. If this is set, then the program exits before running the pipeline. This
means just the validation step and optionally the (if '--plot_graph') pipeline visualization step is run.